Rollout value MP field lookup
